--- .travis.yml
@@ -1,14 +1,0 @@
-language: erlang
-sudo: required
-eng:
-  - ARCH=x64
-  - ARCH=arm
-
-after_success:
-  - bash <(curl -s https://codecov.io/bash)
-
-branches:
-  only:
-    - master
-    - samantha-dev
-    - isaac-dev
