--- gui.py
@@ -1,6 +1,0 @@
-# import packages
-from __future__ import print_function
-from camera import CameraThread
-
-gui = CameraThread(vidStream)
-gui.root.mainloop
